requirements.txt
antlr4-python3-runtime==4.9.3
hydra-core==1.2.0
importlib-resources==5.8.0
omegaconf==2.2.2
packaging==21.3
pyparsing==3.0.9
python-box==6.0.2
python-dotenv==0.20.0
PyYAML==6.0
zipp==3.8.0
